saved
Volume_dipole1-3 will allow you to place a dipole somewhere in the brain and to alter its orientation. The first script sets up the file structures. The second script will simply give you three dipoles at the same location with 3 different orientations. You can alter the values of location and orientation. The third script will allow you to display the dipoles with their corresponding scalp topographies, just as in the .avi files. The scripts Surface_sensitivity1-2 set up the file structures and create a sensitivity map from a standard 128 channel cap and display the result on the cortex, exactly like in the paper.